Trendix/requirements.txt
streamlit>=1.14.0
python-dotenv
pandas
numpy
plotly
openai>=1.3.5
torch
scikit-learn
matplotlib
yfinance
requests
uvicorn
fastapi
watchdog
xgboost
tensorflow==2.16.1
